+++ CHANGELOG.md
@@ -112,5 +112,7 @@
 * Added logging in the `useInstallShuriken` hook
 
 ## 1.15.6
+* Added a check to see if the post_config function exists and if it does to run it.
+* fixed path parsing from script_path and postinstall
 * Fixed `configure` hanging in `install_file`
 * Fixing Os error 2 the file cannot be found